update_2019-12-26--19:44

+++ commit.sh
@@ -1,9 +1,9 @@
 #!/bin/sh
 
-if [[ -f .*~ || -f *~ ]] ; then
-  rm .*~
-  rm *~
-fi
+test -f .*~ 
+[ if $? = "0" ] && rm *~
+test -f *~
+[ if $? = "0" ] && rm .*~
 
 git add --all .
